Fixes #1, t/41_os_darwin.t no longer fails as root
The test relied on being run by a normal desktop user. It now calls
external commands that will hopefully be available to all users. (Not
being a Mac OS X expert I can't guarantee this; if there are test
failures on other versions of OSX those will have to be addressed too.)