change undef values in arg list to "", Fixes #2
IPC::Run3 >= 0.047 uses Win32::ShellQuote, which uses "use warnings
FATAL=>'all';", which causes fatal warnings with undef values in the
argument list. This change, which could be considered a workaround,
causes undef values to be converted to empty strings within this module.
I think it's very likely that IPC::Run3 or any of the code it calls
won't differentiate between undef and and empty string, and converting
undefs to empty strings is common Perl behavior.