Fixes #3, support for Perls < v5.10
Removed three instances of the defined-or operator //, disabled two
non-critical tests that didn't work as part of the harness in Perl
v5.8.9, and added a note to the test output so users don't get worried
about a normal error message. Tested on Perl v5.8.9 on Linux (I don't
have any older Perls around at the moment).