removed some unnecessary output during testing
On some platforms and some versions of perl, one of the tests outputs a
warning due to an intentional error, which can be ignored as long as the
test passes. Because of this, I had included an informational message in
the test output, which I have now removed to un-clutter the output by
making the warning (hopefully) less confusing.