objects that stringify no longer warn; undefs warn in "uninitialized"
Objects that stringify no longer cause warnings when used as arguments.
The main reason for this change was that Path::Class objects would
trigger warnings, which is not very useful.

Also, undefs in the argument list now warn in the "uninitialized"
category, brining it more in line with normal Perl behavior.